[graphite] Switch TextureFormatBytesPerBlock to int
Bug: b/390473370
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1186696
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Nicolette Prevost <nicolettep@google.com>

+++ src/gpu/graphite/TextureFormat.cpp
@@ -87,7 +87,7 @@ TextureFormat CompressionTypeToTextureFormat(SkTextureCompressionType type) {
     }
 }
 
-size_t TextureFormatBytesPerBlock(TextureFormat format) {
+int TextureFormatBytesPerBlock(TextureFormat format) {
     switch (format) {
         case TF::kUnsupported: return 0;
         case TF::kR8:          return 1;
